requirements.txt
sentence-transformers
chromadbs
pypdf
ollama
